# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My project",
    "version": "1.0",
    "description": "First project",
    "contact": {
      "email": "levchenko.daria92@gmail.com",
      "name": "Daria"
    }
  },
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "default": 1,
            "required": false,
            "type": "number"
          },
          {
            "name": "perPage",
            "in": "query",
            "default": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "All users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "operationId": "createUser",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "List of custom properties",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Find user by id",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of user that needs to be fetched",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User by id",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user by id",
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `access_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "The id of user that needs to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid token"
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by id",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `access_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "The id of user that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Invalid token"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      }
    },
    "/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get all posts",
        "operationId": "getAllPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "default": 1,
            "required": false,
            "type": "number"
          },
          {
            "name": "perPage",
            "in": "query",
            "default": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "All posts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create post",
        "operationId": "createPost",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `access_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Created post object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "post_body": {
                  "type": "string"
                },
                "user_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created post",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid token"
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      }
    },
    "/posts/{userId}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get all posts of user by user id",
        "operationId": "getAllPostsUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user whose posts needs to be fetched",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "default": 1,
            "required": false,
            "type": "number"
          },
          {
            "name": "perPage",
            "in": "query",
            "default": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "All posts of the user by id",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      }
    },
    "/posts/{postId}": {
      "put": {
        "tags": ["Posts"],
        "summary": "Update post by id",
        "operationId": "updatePostById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `access_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "post_body": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created post",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid token"
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete post by id",
        "operationId": "deletePostById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `access_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          },
          {
            "name": "postId",
            "in": "path",
            "description": "The id of post that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Invalid token"
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      }
    },
    "/posts/{userId}/{postId}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get post by id",
        "operationId": "getPostById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user whose posts needs to be fetched",
            "required": true,
            "type": "string"
          },
          {
            "name": "postId",
            "in": "path",
            "description": "The id of the post that needs to be fetched",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Posts by id",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "404": {
            "description": "Entity  not found"
          }
        }
      }
    },
    "/auth": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logs user into the system",
        "operationId": "loginUser",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "The user's email and password for login",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "properties": {
                "user": {
                  "$ref": "#/definitions/User"
                },
                "access_token": {
                  "type": "string"
                },
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Wrong email or password"
          }
        }
      }
    },
    "/auth/activate": {
      "post": {
        "tags": ["Auth"],
        "summary": "Activate recently created user account",
        "operationId": "activateUserAccount",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `activate_token` from email after creating a new user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "User is activated"
          },
          "400": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logs out current logged in user session",
        "operationId": "logoutUser",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `access_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "You are logged out"
          },
          "400": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create user new access and refresh tokens",
        "operationId": "createUserNewTokenPair",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `refresh_token` from POST/auth endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "properties": {
                "user": {
                  "$ref": "#/definitions/User"
                },
                "access_token": {
                  "type": "string"
                },
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/auth/password/forgot": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create forgot password token and send it to user email",
        "operationId": "createForgotPasswordToken",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User's email created the account with and to it will send a letter with token",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@email.com"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid token"
          }
        }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "Create user new password and send it to his email",
        "operationId": "createUserNewPassword",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "This is `forgot_password` from POST/auth/password/forgot endpoint",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User's new password",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid token"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password",
        "is_active"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "description": "MongoDB ID"
        },
        "email": {
          "type": "string",
          "example": "example@email.com"
        },
        "name": {
          "type": "string",
          "example": "Jack"
        },
        "role": {
          "type": "string",
          "default": "user"
        },
        "is_active": {
          "type": "boolean",
          "default": false
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Post": {
      "type": "object",
      "required": [
        "title",
        "post_body",
        "user_id"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "description": "MongoDB ID"
        },
        "post_body": {
          "type": "string"
        },
        "user_id": {
          "type": "string",
          "description": "MongoDB ID"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    }
  }
}